Use short build directory
The Node.js build fails when a the directory where it runs has a very
long name:

  execvp: printf: Argument list too long

The reason for that is that the build process generates command line
arguments with very long lists of files, and the total size exceeds
the 128 KiB that the Linux kernel imposes for a single command line
argument:

  http://lxr.free-electrons.com/source/include/uapi/linux/binfmts.h#L14

This limit can't be changed (well, it can be, but requires to change
and recompile the Kernel) so the only thing that we can do is try to
use a shorted build directory. As the build runs inside a mock chroot
that is discarded after the build, we can safely create a short
directory in the root of the filesystem.

Change-Id: Ia290810d033858a1d5ac0ccc94ddaf07ba0c2578
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>